+++ platform/android/SCsub
@@ -4,6 +4,7 @@ Import("env")
 
 android_files = [
     "os_android.cpp",
+    "android_input_handler.cpp",
     "file_access_android.cpp",
     "audio_driver_opensl.cpp",
     "dir_access_jandroid.cpp",
