+++ platform/android/java/lib/src/org/godotengine/godot/input/GodotGestureHandler.java
@@ -75,7 +75,7 @@ public class GodotGestureHandler extends GestureDetector.SimpleOnGestureListener
 		final int x = Math.round(event.getX());
 		final int y = Math.round(event.getY());
 		final int buttonMask = event.getButtonState();
-		queueEvent(() -> GodotLib.doubleTap(buttonMask, x, y));
+		GodotLib.doubleTap(buttonMask, x, y);
 		return true;
 	}
 
@@ -84,7 +84,7 @@ public class GodotGestureHandler extends GestureDetector.SimpleOnGestureListener
 		//Log.i("GodotGesture", "onScroll");
 		final int x = Math.round(distanceX);
 		final int y = Math.round(distanceY);
-		queueEvent(() -> GodotLib.scroll(x, y));
+		GodotLib.scroll(x, y);
 		return true;
 	}
